Support build and deployment within K8 against custom subdomain (#292)
* Add Docker static site containerisation

This change adds Docker containerisation utilising a multi-stage build
Dockerfile. This leads to the static site being generated and served
through nginx.

* Add GHA to build and publish to Quay

This change adds a new GitHub Action that builds the resultant Docker
image and publishes this to the Quay repository.

* Add Helm chart for deployment

This change adds the necessary Helm configuration to deploy the
application within an appropriate Kubernetes cluster.

* Add Drone deployment

This change adds Drone configuration to trigger a new deployment to the
appropriate Kubernetes namespace.