#364 - Review SEGAS-00003 Minimal documentation set for a product
After discussion in the ways of working guild meeting on 7th March 2024 we agreed that a lot of the incident management information is included in the product description section. This is now called out in that section, and the incident management has been rewritten to be about runbooks for expected tasks.